Refactor: Move JSON serialization to entity classes
