+++ backend/src/main/java/com/odde/doughnut/controllers/NotebookController.java
@@ -8,7 +8,6 @@ import com.odde.doughnut.entities.*;
 import com.odde.doughnut.exceptions.UnexpectedNoAccessRightException;
 import com.odde.doughnut.factoryServices.ModelFactoryService;
 import com.odde.doughnut.models.BazaarModel;
-import com.odde.doughnut.models.JsonViewer;
 import com.odde.doughnut.models.UserModel;
 import com.odde.doughnut.services.NotebookIndexingService;
 import com.odde.doughnut.services.ObsidianFormatService;
@@ -62,7 +61,7 @@ class NotebookController {
 
     User user = currentUser.getEntity();
     NotebooksViewedByUser notebooksViewedByUser =
-        new JsonViewer().jsonNotebooksViewedByUser(user.getOwnership().getNotebooks());
+        user.getOwnership().jsonNotebooksViewedByUser(user.getOwnership().getNotebooks());
     notebooksViewedByUser.subscriptions = user.getSubscriptions();
     return notebooksViewedByUser;
   }
